summarizer/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-adk